Move _ft_func related items into utils.py along with the enum code.
